requirements.txt
# YouTube and Google Sheets API dependencies
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0
python-dotenv>=1.0.0
